move CommerceBindings type to ts file
https://coveord.atlassian.net/browse/KIT-3941
